Move classes related to cut operators to a separate header file (#1525)
